Move subtitle files to Subtitle folder
